chore: remove rust/kona/tests/supervisor and the sysgo op-supervisor path
The rust/kona/tests/supervisor packages are not invoked by any CI job
(the kona test-e2e-sysgo matrix only runs node/common, node/reorgs and
node/restart). They were the last consumer of the
NewSimpleInterop/NewSingleChainInterop presets, which in turn were the
last code path that booted op-supervisor as a service in devstack.

With those tests gone:
- Delete NewSimpleInterop / NewSingleChainInterop factories and the
  singleChainInteropFromRuntime / simpleInteropFromRuntime adapters.
- Delete NewSimpleInteropRuntime{,WithConfig} /
  NewSingleChainInteropRuntime{,WithConfig} plus the startSupervisor,
  startOPSupervisor, startKonaSupervisor, OpSupervisor, KonaSupervisor,
  and Supervisor-interface chain.
- Drop MultiChainRuntime.PrimarySupervisor and
  SingleChainInteropSupport.Supervisor (no setters/readers left).
- Drop SingleChainInterop.Supervisor, the dsl.Supervisor wrapper,
  supervisorFrontend, newSupervisorFrontend, newPresetSupervisor, and
  the stack.Supervisor interface.
- Move readJWTSecretFromPath and validateSimpleInteropPresetConfig to
  a small interop_helpers.go since they are still needed by the
  supernode runtime.
