+++ op-devstack/presets/superproofs_from_runtime.go
@@ -38,7 +38,6 @@ func simpleInteropFromSupernodeProofsRuntime(t devtest.T, runtime *sysgo.MultiCh
 			Log:              t.Logger(),
 			T:                t,
 			timeTravel:       runtime.TimeTravel,
-			Supervisor:       nil,
 			SuperRoots:       dsl.NewSupernodeWithTestControl(supernodeFrontend, runtime.Supernode),
 			TestSequencer:    dsl.NewTestSequencer(testSequencer),
 			L1Network:        twoL2.L1Network,
@@ -133,7 +132,6 @@ func singleChainInteropFromSupernodeProofsRuntime(t devtest.T, runtime *sysgo.Mu
 		Log:              t.Logger(),
 		T:                t,
 		timeTravel:       runtime.TimeTravel,
-		Supervisor:       nil,
 		SuperRoots:       dsl.NewSupernodeWithTestControl(supernodeFrontend, runtime.Supernode),
 		TestSequencer:    dsl.NewTestSequencer(testSequencer),
 		L1Network:        dsl.NewL1Network(l1Network, l1ELDSL, l1CLDSL),
