--- op-devstack/sysgo/supervisor.go
@@ -1,8 +1,0 @@
-package sysgo
-
-import "github.com/ethereum-optimism/optimism/op-devstack/stack"
-
-type Supervisor interface {
-	stack.Lifecycle
-	UserRPC() string
-}
